scripts/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
requests==2.31.0
beautifulsoup4==4.12.2
groq==0.4.2
textblob==0.17.1
psycopg2-binary==2.9.7
python-dotenv==1.0.0
lxml==4.9.3
selenium==4.15.2
webdriver-manager==4.0.1